chore: renamed `eth` to `evm`
In general all methods that we're using for eth are also appliable for evm
