+++ omni-relayer/config.toml
@@ -6,7 +6,7 @@ network = "testnet"
 rpc_url = "https://rpc.testnet.near.org/"
 token_locker_id = "omni-locker.testnet"
 
-[eth]
+[evm]
 rpc_http_url = "https://eth-sepolia.g.alchemy.com/v2/API-KEY"
 rpc_ws_url = "wss://eth-sepolia.g.alchemy.com/v2/API-KEY"
 chain_id = 11_155_111
