+++ omni-relayer/src/startup/mod.rs
@@ -6,7 +6,7 @@ use omni_connector::{OmniConnector, OmniConnectorBuilder};
 
 use crate::config;
 
-pub mod eth;
+pub mod evm;
 pub mod near;
 
 pub fn build_connector(
@@ -16,11 +16,11 @@ pub fn build_connector(
     info!("Building Omni connector");
 
     OmniConnectorBuilder::default()
-        .eth_endpoint(Some(config.eth.rpc_http_url.clone()))
-        .eth_chain_id(Some(config.eth.chain_id))
+        .eth_endpoint(Some(config.evm.rpc_http_url.clone()))
+        .eth_chain_id(Some(config.evm.chain_id))
         .near_endpoint(Some(config.near.rpc_url.clone()))
         .token_locker_id(Some(config.near.token_locker_id.to_string()))
-        .bridge_token_factory_address(Some(config.eth.bridge_token_factory_address.to_string()))
+        .bridge_token_factory_address(Some(config.evm.bridge_token_factory_address.to_string()))
         .eth_private_key(Some(
             std::env::var("ETH_PRIVATE_KEY")
                 .context("Failed to get `ETH_PRIVATE_KEY` env variable")?,
