+++ omni-relayer/src/workers/mod.rs
@@ -1,2 +1,2 @@
-pub mod eth;
+pub mod evm;
 pub mod near;
